Trim BGE-M3 pooling surface
